moved history to dumb component folder
